Conflicts:
	.gitignore
	batch-shell/dist/shells/cajamar/apr_wait_alert.sh
	rec-common/src/main/java/es/capgemini/pfs/acuerdo/dao/AcuerdoDao.java
	rec-common/src/main/java/es/capgemini/pfs/acuerdo/dao/impl/AcuerdoDaoImpl.java
	rec-web/src/main/webapp/WEB-INF/config/appMessages.properties
	recovery-busquedaProcedimientos-plugin/src/main/java/es/pfsgroup/plugin/recovery/busquedaProcedimientos/procedimiento/dao/impl/BPRProcedimientoDaoImpl.java
	recovery-config-plugin/src/main/resources/plugins/messages/plugin-config-messages.properties
	recovery-config-plugin/src/web/flows/plugin/config/despachoExterno/ADMmodificarDespachoExterno.xml
	recovery-config-plugin/src/web/jsp/plugin/config/despachoExterno/altaDespachoExterno.jsp
	recovery-coreextension-plugin/src/main/java/es/pfsgroup/plugin/recovery/coreextension/coreextensionController.java
	recovery-coreextension-plugin/src/main/java/es/pfsgroup/recovery/ext/impl/asunto/dto/EXTDtoBusquedaAsunto.java
	recovery-coreextension-plugin/src/main/resources/plugins/flows/asunto/listadoBusquedaAsuntos.xml
	recovery-itinerarios-plugin/src/main/resources/plugins/messages/plugin-itinerarios-messages.properties
	recovery-nuevoModeloBienes-plugin/src/main/resources/plugins/messages/plugin-nuevoModeloBienes-messages.properties
	recovery-precontencioso-plugin/src/main/java/es/pfsgroup/plugin/precontencioso/documento/controller/DocumentoPCOController.java
	recovery-precontencioso-plugin/src/main/java/es/pfsgroup/plugin/precontencioso/documento/dto/DocumentoPCODto.java
	recovery-precontencioso-plugin/src/main/java/es/pfsgroup/plugin/precontencioso/documento/model/DocumentoPCO.java
	recovery-precontencioso-plugin/src/main/java/es/pfsgroup/plugin/precontencioso/expedienteJudicial/dao/impl/ProcedimientoPCODaoImpl.java
	recovery-precontencioso-plugin/src/main/java/es/pfsgroup/plugin/precontencioso/liquidacion/manager/LiquidacionManager.java
	recovery-precontencioso-plugin/src/web/jsp/plugin/precontencioso/documento/popups/editarDocumento.jsp
	sql/9.1/cajamar/informacional_old/ddl/DDL_000_RECOVERY_CM_DWH_CARGAR_H_BIEN.sql
	sql/9.2/bankia/DML_101_ENTITY01_INSERT_MIG_DEE_DESPACHO_EXTRAS_TO_DEE_DESPACHO_EXTRAS.sql

Former-commit-id: d83d142839227b4ab1c753512fe68c7e6af39372

+++ recovery-coreextension-plugin/src/main/java/es/pfsgroup/plugin/recovery/nuevoModeloBienes/model/NMBBien.java
@@ -1,5 +1,6 @@
 package es.pfsgroup.plugin.recovery.nuevoModeloBienes.model;
 
+import java.math.BigDecimal;
 import java.math.BigInteger;
 import java.util.Date;
 import java.util.List;
@@ -154,7 +155,7 @@ public class NMBBien extends Bien implements NMBBienInfo{
 	private Boolean obraEnCurso;
 	
 	@Column(name= "BIE_TIPO_SUBASTA")
-	private Float tipoSubasta;
+	private BigDecimal tipoSubasta;
 	
 	@Column(name = "BIE_DUE_DILLIGENCE")
 	private Boolean dueDilligence;
@@ -243,11 +244,11 @@ public class NMBBien extends Bien implements NMBBienInfo{
 //    private String	tareaActivaLlaves;
 	
 	
-	public Float getTipoSubasta() {
+	public BigDecimal getTipoSubasta() {
 		return tipoSubasta;
 	}
 
-	public void setTipoSubasta(Float tipoSubasta) {
+	public void setTipoSubasta(BigDecimal tipoSubasta) {
 		this.tipoSubasta = tipoSubasta;
 	}
